DOMA-297 move ExportToExcelActionBar to common
